requirements.txt
# Apache Airflow
apache-airflow==2.8.0

# Data Processing
pyspark==3.5.0
pyarrow==14.0.2

# API Requests
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Data Quality
soda-core-spark==3.3.2

# Utilities
python-dotenv==1.0.0
